+++ kconfig/parser/expr.c
@@ -1191,7 +1191,7 @@ static void expr_print_gstr_helper(void *data, struct symbol *sym, const char *s
 
 		last_line_length = strlen(gs->s) - (last_cr - gs->s);
 
-		if ((last_line_length + extra_length) > gs->max_width)
+		if ((int)(last_line_length + extra_length) > gs->max_width)
 			str_append(gs, "\\\n");
 	}
